# docs/swagger.json
{
	"components": {
...
		"securitySchemes": {}
	},
	"info": {
		"title": "express_typescript_skeleton",
		"version": "1.0.0",
		"description": "a basic express app writen in typescript",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Abdulraheem A omeiza"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/api/auth/google/google-url": {
			"get": {
				"operationId": "GetGoogleUrl",
				"responses": {
					"201": {
						"description": "link generated successfully",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"details": {},
												"message": {
													"type": "string"
												},
												"status": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"details",
												"message",
												"status"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"details": {},
												"message": {
													"type": "string"
												},
												"status": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"details",
												"message",
												"status"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "user not found"
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"details": {},
												"message": {
													"type": "string"
												},
												"status": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"details",
												"message",
												"status"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [
					{
						"description": "- either login or registration",
						"in": "query",
						"name": "action",
						"required": true,
						"schema": {
							"type": "string",
							"enum": [
								"login",
								"register"
							]
						}
					}
				]
			}
		},
		"/api/auth/google/callback": {
			"get": {
				"operationId": "GetGoogleDetails",
				"responses": {
					"201": {
						"description": "successfully",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"details": {},
												"message": {
													"type": "string"
												},
												"status": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"details",
												"message",
												"status"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"details": {},
												"message": {
													"type": "string"
												},
												"status": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"details",
												"message",
												"status"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "user not found"
					},
					"422": {
						"description": "validation failed"
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"details": {},
												"message": {
													"type": "string"
												},
												"status": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"details",
												"message",
												"status"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [
					{
						"description": "- this is the code sent by google",
						"in": "query",
						"name": "code",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "- either login or registration link",
						"in": "query",
						"name": "action",
						"required": true,
						"schema": {}
					}
				]
			}
		},
		"/api/auth/google/validate-token": {
			"post": {
				"operationId": "GetGoogleToken",
				"responses": {
					"201": {
						"description": "successfully",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"details": {},
												"message": {
													"type": "string"
												},
												"status": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"details",
												"message",
												"status"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"details": {},
												"message": {
													"type": "string"
												},
												"status": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"details",
												"message",
												"status"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"resp": {
											"properties": {
												"details": {},
												"message": {
													"type": "string"
												},
												"status": {
													"type": "boolean",
													"enum": [
														true,
														false
													]
												}
											},
											"required": [
												"details",
												"message",
												"status"
											],
											"type": "object"
										}
									},
									"required": [
										"resp"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [
					{
						"description": "- header authorization token",
						"in": "header",
						"name": "Authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}